Modified box.py to call haplaylist shell script
Box.py now calls haplaylist.sh and passes the plist variable. The haplaylist.sh calls a curl command to a home assistant instance running on another machine and passes the playlist variable to an input_text component. Also included is the musiccards.service for systemd to start box.py automatically on boot.